AI water/requirements.txt
langchain
llama-index
pydantic
python-dotenv
fastapi
uvicorn
streamlit
typer #CLI integration
sqlalchemy
langchain-google-genai